Boundaries
Boundaries where seperated into functions to increase the legibility of the code at the expence of some efficeny I thought this was nessessary to stop the code from becoming crowded. Set_integration_boundaries sets the lower and upper bounds of the integration for loop baised on the placment of the processor submatrix. Current_velocity was also put into a matrix with inputs the i and j of the for loop. This functions was used to input the values that where recived from other processors to use in integrations