Assembly
Created an Assembly matrix this was used to print out the results by adding up all the procssessors to a matrix of Nx by Ny. Then a print function was creted to print to a text file. This could then be used in matlab to plot the results. In addition the make file is now changed to be able to run single, parallel and px py